+++ chess-app/src/engine/miniMax/index.ts
@@ -1,0 +1,3 @@
+export * from "./Engine"
+export * from "./evalConstants"
+export * from "./EvaluateBoard"
